requirements.txt
requests
pandas
pytest
python-dotenv
pyodbc
setuptools
flask
